remove predocs.py
