+++ 5.internal-document-search/scripts/prepdocs.py
@@ -1,249 +1,157 @@
-import os
 import argparse
-import glob
-import html
-import io
-import re
-import time
-from pypdf import PdfReader, PdfWriter
 from azure.identity import AzureDeveloperCliCredential
 from azure.identity import ManagedIdentityCredential
 from azure.core.credentials import AzureKeyCredential
-from azure.storage.blob import BlobServiceClient
 from azure.search.documents.indexes import SearchIndexClient
 from azure.search.documents.indexes.models import *
-from azure.search.documents import SearchClient
 from azure.ai.formrecognizer import DocumentAnalysisClient
+from azure.search.documents import SearchClient
+from azure.storage.blob import BlobServiceClient
+from pypdf import PdfReader, PdfWriter
+from glob import glob
+import io
+import os
+import re
 
-MAX_SECTION_LENGTH = 1000
-SENTENCE_SEARCH_LIMIT = 100
-SECTION_OVERLAP = 100
+from typing import Dict, List, Tuple
 
-parser = argparse.ArgumentParser(
-    description="Prepare documents by extracting content from PDFs, splitting content into sections, uploading to blob storage, and indexing in a search index.",
-    epilog="Example: prepdocs.py '..\data\*' --storageaccount myaccount --container mycontainer --searchservice mysearch --index myindex -v"
-    )
-parser.add_argument("files", help="Files to be processed")
-parser.add_argument("--category", help="Value for the category field in the search index for all sections indexed in this run")
-parser.add_argument("--skipblobs", action="store_true", help="Skip uploading individual pages to Azure Blob Storage")
-parser.add_argument("--storageaccount", help="Azure Blob Storage account name")
-parser.add_argument("--container", help="Azure Blob Storage container name")
-parser.add_argument("--storagekey", required=False, help="Optional. Use this Azure Blob Storage account key instead of the current user identity to login (use az login to set current user for Azure)")
-parser.add_argument("--tenantid", required=False, help="Optional. Use this to define the Azure directory where to authenticate)")
-parser.add_argument("--searchservice", help="Name of the Azure Cognitive Search service where content should be indexed (must exist already)")
-parser.add_argument("--index", help="Name of the Azure Cognitive Search index where content should be indexed (will be created if it doesn't exist)")
-parser.add_argument("--searchkey", required=False, help="Optional. Use this Azure Cognitive Search account key instead of the current user identity to login (use az login to set current user for Azure)")
-parser.add_argument("--remove", action="store_true", help="Remove references to this document from blob storage and the search index")
-parser.add_argument("--removeall", action="store_true", help="Remove all blobs from blob storage and documents from the search index")
-parser.add_argument("--localpdfparser", action="store_true", help="Use PyPdf local PDF parser (supports only digital PDFs) instead of Azure Form Recognizer service to extract text, tables and layout from the documents")
-parser.add_argument("--formrecognizerservice", required=False, help="Optional. Name of the Azure Form Recognizer service which will be used to extract text, tables and layout from the documents (must exist already)")
-parser.add_argument("--formrecognizerkey", required=False, help="Optional. Use this Azure Form Recognizer account key instead of the current user identity to login (use az login to set current user for Azure)")
-parser.add_argument("--verbose", "-v", action="store_true", help="Verbose output")
-parser.add_argument('--managedidentitycredential', action='store_true', help='Use Managed Identity (e.g., Cloud Shell) credentials')
-args = parser.parse_args()
 
-# Use the current user identity to connect to Azure services unless a key is explicitly set for any of them
-if args.managedidentitycredential:
-    default_creds = ManagedIdentityCredential()
-else:
-    azd_credential = AzureDeveloperCliCredential() if args.tenantid is None else AzureDeveloperCliCredential(tenant_id=args.tenantid, process_timeout=60)
-    default_creds = azd_credential if args.searchkey is None or args.storagekey is None else None
-search_creds = default_creds if args.searchkey is None else AzureKeyCredential(args.searchkey)
+# ログ出力用のクラス
+class Logger:
+	def __init__(self, verbose: bool):
+		self.verbose = verbose
 
-if not args.skipblobs:
-    storage_creds = default_creds if args.storagekey is None else args.storagekey
-if not args.localpdfparser:
-    # check if Azure Form Recognizer credentials are provided
-    if args.formrecognizerservice is None:
+	def log(self, logging_message: str):
+		if self.verbose:
+			print(logging_message)
+
+
+# 引数を読み込む関数
+def read_argument() -> Tuple[Dict[str, str], Logger]:
+	parser = argparse.ArgumentParser(
+	description="Prepare documents by extracting content from PDFs, splitting content into sections, uploading to blob storage, and indexing in a search index.",
+	epilog="Example: prepdocs.py '..\data\*' --storageaccount myaccount --container mycontainer --searchservice mysearch --index myindex -v"
+	)
+	parser.add_argument("files", help="Files to be processed")
+	parser.add_argument("--storageaccount", help="Azure Blob Storage account name")
+	parser.add_argument("--storagekey", required=False, help="Optional. Use this Azure Blob Storage account key instead of the current user identity to login (use az login to set current user for Azure)")
+	parser.add_argument("--container", help="Azure Blob Storage container name")
+	parser.add_argument("--tenantid", required=False, help="Optional. Use this to define the Azure directory where to authenticate)")
+	parser.add_argument("--searchservice", help="Name of the Azure Cognitive Search service where content should be indexed (must exist already)")
+	parser.add_argument("--searchkey", required=False, help="Optional. Use this Azure Cognitive Search account key instead of the current user identity to login (use az login to set current user for Azure)")
+	parser.add_argument("--index", help="Name of the Azure Cognitive Search index where content should be indexed (will be created if it doesn't exist)")
+	parser.add_argument("--formrecognizerservice", required=False, help="Optional. Name of the Azure Form Recognizer service which will be used to extract text, tables and layout from the documents (must exist already)")
+	parser.add_argument("--formrecognizerkey", required=False, help="Optional. Use this Azure Form Recognizer account key instead of the current user identity to login (use az login to set current user for Azure)")
+	parser.add_argument("--verbose", "-v", action="store_true", help="Verbose output")
+	parser.add_argument('--managedidentitycredential', action='store_true', help='Use Managed Identity (e.g., Cloud Shell) credentials')
+	args = parser.parse_args()
+
+	logger = Logger(args.verbose)
+
+	return args, logger
+
+
+# Azure Developer CLIの認証情報を取得する関数
+def _get_azd_credential(tenantid: str):
+	if tenantid is None:
+		return AzureDeveloperCliCredential()
+	else:
+         return AzureDeveloperCliCredential(tenant_id=tenantid, process_timeout=60)
+
+
+# 認証方法を決定する関数
+def init_default_credential(managedidentitycredential: bool, searchkey: str, storagekey: str, tenantid: str):
+	# --managedidentitycredentialオプションが指定されたとき
+	if managedidentitycredential:
+		return ManagedIdentityCredential()
+	# keyが指定されないとき
+	if searchkey is None or storagekey is None:
+		azd_credential = _get_azd_credential(tenantid)
+		return azd_credential
+	# keyが指定されたとき
+	else:
+		return None
+
+
+# Search Serviceの認証情報を取得する関数
+def init_search_service_credential(service_name: str, secret_key: str, default_creds):
+    if service_name is None:
+        print("Error: Search service is not provided.")
+        exit(1)
+    if secret_key is None:
+        return default_creds
+    else:
+        return AzureKeyCredential(secret_key)
+
+
+# Form Recognizerの認証情報を取得する関数
+def init_form_recognizer_credential(service_name: str, secret_key: str, default_creds):
+    if service_name is None:
         print("Error: Azure Form Recognizer service is not provided. Please provide formrecognizerservice or use --localpdfparser for local pypdf parser.")
         exit(1)
-    formrecognizer_creds = default_creds if args.formrecognizerkey is None else AzureKeyCredential(args.formrecognizerkey)
-
-def blob_name_from_file_page(filename, page = 0):
-    if os.path.splitext(filename)[1].lower() == ".pdf":
-        return os.path.splitext(os.path.basename(filename))[0] + f"-{page}" + ".pdf"
+    if secret_key is None:
+        return default_creds
     else:
-        return os.path.basename(filename)
+        return AzureKeyCredential(secret_key)
 
-def upload_blobs(filename):
-    blob_service = BlobServiceClient(account_url=f"https://{args.storageaccount}.blob.core.windows.net", credential=storage_creds)
-    blob_container = blob_service.get_container_client(args.container)
-    if not blob_container.exists():
-        blob_container.create_container()
 
-    # if file is PDF split into pages and upload each page as a separate blob
-    if os.path.splitext(filename)[1].lower() == ".pdf":
-        reader = PdfReader(filename)
-        pages = reader.pages
-        for i in range(len(pages)):
-            blob_name = blob_name_from_file_page(filename, i)
-            if args.verbose: print(f"\tUploading blob for page {i} -> {blob_name}")
-            f = io.BytesIO()
-            writer = PdfWriter()
-            writer.add_page(pages[i])
-            writer.write(f)
-            f.seek(0)
-            blob_container.upload_blob(blob_name, f, overwrite=True)
-    else:
-        blob_name = blob_name_from_file_page(filename)
-        with open(filename,"rb") as data:
-            blob_container.upload_blob(blob_name, data, overwrite=True)
+# Storage Accountの認証情報を取得する関数
+def init_storage_credential(storage_account_name: str, secret_key: str, default_creds):
+	if storage_account_name is None:
+		print("Error: Storage account is not provided.")
+		exit(1)
+	if secret_key is None:
+		return default_creds
+	else:
+		return secret_key
 
-def remove_blobs(filename):
-    if args.verbose: print(f"Removing blobs for '{filename or '<all>'}'")
-    blob_service = BlobServiceClient(account_url=f"https://{args.storageaccount}.blob.core.windows.net", credential=storage_creds)
-    blob_container = blob_service.get_container_client(args.container)
-    if blob_container.exists():
-        if filename is None:
-            blobs = blob_container.list_blob_names()
-        else:
-            prefix = os.path.splitext(os.path.basename(filename))[0]
-            blobs = filter(lambda b: re.match(f"{prefix}-\d+\.pdf", b), blob_container.list_blob_names(name_starts_with=os.path.splitext(os.path.basename(prefix))[0]))
-        for b in blobs:
-            if args.verbose: print(f"\tRemoving blob {b}")
-            blob_container.delete_blob(b)
 
-def table_to_html(table):
-    table_html = "<table>"
-    rows = [sorted([cell for cell in table.cells if cell.row_index == i], key=lambda cell: cell.column_index) for i in range(table.row_count)]
-    for row_cells in rows:
-        table_html += "<tr>"
-        for cell in row_cells:
-            tag = "th" if (cell.kind == "columnHeader" or cell.kind == "rowHeader") else "td"
-            cell_spans = ""
-            if cell.column_span > 1: cell_spans += f" colSpan={cell.column_span}"
-            if cell.row_span > 1: cell_spans += f" rowSpan={cell.row_span}"
-            table_html += f"<{tag}{cell_spans}>{html.escape(cell.content)}</{tag}>"
-        table_html +="</tr>"
-    table_html += "</table>"
-    return table_html
+# PDFファイルからテキストを抽出する関数
+def recognize_text_from_pdf(target_file_path: str, form_recognizer_name: str, form_recognizer_creds) -> List[Dict[str, str]]:
+	form_recognizer_client = DocumentAnalysisClient(
+		endpoint=f"https://{form_recognizer_name}.cognitiveservices.azure.com/",
+		credential=form_recognizer_creds,
+		headers={"x-ms-useragent": "azure-search-chat-demo/1.0.0"}
+	)
+	# ファイルを読み込み、テキストを抽出
+	with open(target_file_path, "rb") as f:
+		poller = form_recognizer_client.begin_analyze_document("prebuilt-layout", document = f)
+	form_recognizer_results = poller.result()
+	# ページごとにテキストを結合
+	document_content_list = []
+	target_file_name = os.path.basename(target_file_path)
+	for page in form_recognizer_results.pages:
+		page_content = "\n".join([line.content for line in page.lines])
+		document_content_list.append({
+			"id": re.sub("[^0-9a-zA-Z_-]","_",f"{target_file_name}-{page.page_number}"),
+            "content": page_content,
+            "category": None,
+            "sourcepage": f"{os.path.splitext(target_file_name)[0]}-{page.page_number}.pdf",
+            "sourcefile": target_file_name
+		})
+	return document_content_list
 
-def get_document_text(filename):
-    offset = 0
-    page_map = []
-    if args.localpdfparser:
-        reader = PdfReader(filename)
-        pages = reader.pages
-        for page_num, p in enumerate(pages):
-            page_text = p.extract_text()
-            page_map.append((page_num, offset, page_text))
-            offset += len(page_text)
-    else:
-        if args.verbose: print(f"Extracting text from '{filename}' using Azure Form Recognizer")
-        form_recognizer_client = DocumentAnalysisClient(endpoint=f"https://{args.formrecognizerservice}.cognitiveservices.azure.com/", credential=formrecognizer_creds, headers={"x-ms-useragent": "azure-search-chat-demo/1.0.0"})
-        with open(filename, "rb") as f:
-            poller = form_recognizer_client.begin_analyze_document("prebuilt-layout", document = f)
-        form_recognizer_results = poller.result()
 
-        for page_num, page in enumerate(form_recognizer_results.pages):
-            tables_on_page = [table for table in form_recognizer_results.tables if table.bounding_regions[0].page_number == page_num + 1]
-
-            # mark all positions of the table spans in the page
-            page_offset = page.spans[0].offset
-            page_length = page.spans[0].length
-            table_chars = [-1]*page_length
-            for table_id, table in enumerate(tables_on_page):
-                for span in table.spans:
-                    # replace all table spans with "table_id" in table_chars array
-                    for i in range(span.length):
-                        idx = span.offset - page_offset + i
-                        if idx >=0 and idx < page_length:
-                            table_chars[idx] = table_id
-
-            # build page text by replacing charcters in table spans with table html
-            page_text = ""
-            added_tables = set()
-            for idx, table_id in enumerate(table_chars):
-                if table_id == -1:
-                    page_text += form_recognizer_results.content[page_offset + idx]
-                elif table_id not in added_tables:
-                    page_text += table_to_html(tables_on_page[table_id])
-                    added_tables.add(table_id)
-
-            page_text += " "
-            page_map.append((page_num, offset, page_text))
-            print(page)
-            offset += len(page_text)
-
-    return page_map
-
-def split_text(page_map):
-    SENTENCE_ENDINGS = [".", "!", "?"]
-    WORDS_BREAKS = [",", ";", ":", " ", "(", ")", "[", "]", "{", "}", "\t", "\n"]
-    if args.verbose: print(f"Splitting '{filename}' into sections")
-
-    def find_page(offset):
-        l = len(page_map)
-        for i in range(l - 1):
-            if offset >= page_map[i][1] and offset < page_map[i + 1][1]:
-                return i
-        return l - 1
-
-    all_text = "".join(p[2] for p in page_map)
-    length = len(all_text)
-    start = 0
-    end = length
-    while start + SECTION_OVERLAP < length:
-        last_word = -1
-        end = start + MAX_SECTION_LENGTH
-
-        if end > length:
-            end = length
-        else:
-            # Try to find the end of the sentence
-            while end < length and (end - start - MAX_SECTION_LENGTH) < SENTENCE_SEARCH_LIMIT and all_text[end] not in SENTENCE_ENDINGS:
-                if all_text[end] in WORDS_BREAKS:
-                    last_word = end
-                end += 1
-            if end < length and all_text[end] not in SENTENCE_ENDINGS and last_word > 0:
-                end = last_word # Fall back to at least keeping a whole word
-        if end < length:
-            end += 1
-
-        # Try to find the start of the sentence or at least a whole word boundary
-        last_word = -1
-        while start > 0 and start > end - MAX_SECTION_LENGTH - 2 * SENTENCE_SEARCH_LIMIT and all_text[start] not in SENTENCE_ENDINGS:
-            if all_text[start] in WORDS_BREAKS:
-                last_word = start
-            start -= 1
-        if all_text[start] not in SENTENCE_ENDINGS and last_word > 0:
-            start = last_word
-        if start > 0:
-            start += 1
-
-        section_text = all_text[start:end]
-        yield (section_text, find_page(start))
-
-        last_table_start = section_text.rfind("<table")
-        if (last_table_start > 2 * SENTENCE_SEARCH_LIMIT and last_table_start > section_text.rfind("</table")):
-            # If the section ends with an unclosed table, we need to start the next section with the table.
-            # If table starts inside SENTENCE_SEARCH_LIMIT, we ignore it, as that will cause an infinite loop for tables longer than MAX_SECTION_LENGTH
-            # If last table starts inside SECTION_OVERLAP, keep overlapping
-            if args.verbose: print(f"Section ends with unclosed table, starting next section with the table at page {find_page(start)} offset {start} table start {last_table_start}")
-            start = min(end - SECTION_OVERLAP, start + last_table_start)
-        else:
-            start = end - SECTION_OVERLAP
-        
-    if start + SECTION_OVERLAP < end:
-        yield (all_text[start:end], find_page(start))
-
-def create_sections(filename, page_map):
-    for i, (section, pagenum) in enumerate(split_text(page_map)):
-        yield {
-            "id": re.sub("[^0-9a-zA-Z_-]","_",f"{filename}-{i}"),
-            "content": section,
-            "category": args.category,
-            "sourcepage": blob_name_from_file_page(filename, pagenum),
-            "sourcefile": filename
-        }
-
-def create_search_index():
-    if args.verbose: print(f"Ensuring search index {args.index} exists")
-    index_client = SearchIndexClient(endpoint=f"https://{args.searchservice}.search.windows.net/",
-                                     credential=search_creds)
-    if args.index not in index_client.list_index_names():
-        index = SearchIndex(
-            name=args.index,
+# Search Serviceのインデックスを作成する関数
+def create_search_service_index(search_creds, search_service_name: str, index_name: str):
+	search_service_index_client = SearchIndexClient(
+		endpoint=f"https://{search_service_name}.search.windows.net/",
+		credential=search_creds
+	)
+	# indexが登録済みの場合
+	if index_name in search_service_index_client.list_index_names():
+		print(f"Search index {index_name} already exists")
+	else:
+		semantic_config = SemanticConfiguration(
+			name='default',
+			prioritized_fields=PrioritizedFields(
+				title_field=None,
+				prioritized_content_fields=[SemanticField(field_name='content')]
+			)
+		)
+		# id, content, category, sourcepage, sourcefileの5つのフィールドを持つindexを作成
+		search_service_index = SearchIndex(
+            name=index_name,
             fields=[
                 SimpleField(name="id", type="Edm.String", key=True),
                 SearchableField(name="content", type="Edm.String", analyzer_name="ja.microsoft"),
@@ -252,71 +160,149 @@ def create_search_index():
                 SimpleField(name="sourcefile", type="Edm.String", filterable=True, facetable=True)
             ],
             semantic_settings=SemanticSettings(
-                configurations=[SemanticConfiguration(
-                    name='default',
-                    prioritized_fields=PrioritizedFields(
-                        title_field=None, prioritized_content_fields=[SemanticField(field_name='content')]))])
+                configurations=[semantic_config]
+			)
         )
-        if args.verbose: print(f"Creating {args.index} search index")
-        index_client.create_index(index)
-    else:
-        if args.verbose: print(f"Search index {args.index} already exists")
+		try:
+			search_service_index_client.create_index(search_service_index)
+		except Exception as ex:
+			print(f"Error creating search index {index_name}: {ex}")
+			exit(1)
 
-def index_sections(filename, sections):
-    if args.verbose: print(f"Indexing sections from '{filename}' into search index '{args.index}'")
-    search_client = SearchClient(endpoint=f"https://{args.searchservice}.search.windows.net/",
-                                    index_name=args.index,
-                                    credential=search_creds)
-    i = 0
-    batch = []
-    for s in sections:
-        batch.append(s)
-        i += 1
-        if i % 1000 == 0:
-            results = search_client.upload_documents(documents=batch)
-            succeeded = sum([1 for r in results if r.succeeded])
-            if args.verbose: print(f"\tIndexed {len(results)} sections, {succeeded} succeeded")
-            batch = []
 
-    if len(batch) > 0:
-        results = search_client.upload_documents(documents=batch)
-        succeeded = sum([1 for r in results if r.succeeded])
-        if args.verbose: print(f"\tIndexed {len(results)} sections, {succeeded} succeeded")
+# search indexにドキュメントを追加する関数
+def add_documents_to_index(
+		search_service_name: str,
+		index_name: str,
+		search_creds,
+		document_content_list: List[Dict[str, str]],
+		logger: Logger
+	):
+	search_client = SearchClient(
+		endpoint=f"https://{search_service_name}.search.windows.net/",
+		index_name=index_name,
+		credential=search_creds
+	)
+	# ドキュメントをindexに追加
+	results = search_client.upload_documents(documents=document_content_list)
+	n_success_docs = sum([1 for result in results if result.succeeded])
+	logger.log(f"\tCreate Index Succeeded: {n_success_docs}/{len(results)} Pages")
 
-def remove_from_index(filename):
-    if args.verbose: print(f"Removing sections from '{filename or '<all>'}' from search index '{args.index}'")
-    search_client = SearchClient(endpoint=f"https://{args.searchservice}.search.windows.net/",
-                                    index_name=args.index,
-                                    credential=search_creds)
-    while True:
-        filter = None if filename is None else f"sourcefile eq '{os.path.basename(filename)}'"
-        r = search_client.search("", filter=filter, top=1000, include_total_count=True)
-        if r.get_count() == 0:
-            break
-        r = search_client.delete_documents(documents=[{ "id": d["id"] } for d in r])
-        if args.verbose: print(f"\tRemoved {len(r)} sections from index")
-        # It can take a few seconds for search results to reflect changes, so wait a bit
-        time.sleep(2)
 
-if args.removeall:
-    remove_blobs(None)
-    remove_from_index(None)
-else:
-    if not args.remove:
-        create_search_index()
-    
-    print("Processing files...")
-    for filename in glob.glob(args.files):
-        if args.verbose: print(f"Processing '{filename}'")
-        if args.remove:
-            remove_blobs(filename)
-            remove_from_index(filename)
-        elif args.removeall:
-            remove_blobs(None)
-            remove_from_index(None)
-        else:
-            if not args.skipblobs:
-                upload_blobs(filename)
-            page_map = get_document_text(filename)
-            sections = create_sections(os.path.basename(filename), page_map)
-            index_sections(os.path.basename(filename), sections)
+# ドキュメントをblob storageにアップロードする関数
+def upload_documents_to_blob_storage(
+		blob_service_name: str,
+		container_name: str,
+		target_file_path: str,
+		storage_creds,
+		logger: Logger
+	):
+
+	blob_service_client = BlobServiceClient(
+		account_url=f"https://{blob_service_name}.blob.core.windows.net",
+		credential=storage_creds
+	)
+	container_client = blob_service_client.get_container_client(container_name)
+	target_file_name = os.path.basename(target_file_path)
+	# コンテナが存在しない場合は作成
+	if not container_client.exists():
+		container_client.create_container()
+
+	# PDFファイルの場合
+	if os.path.splitext(target_file_name)[1].lower() == ".pdf":
+		reader = PdfReader(target_file_path)
+		pages = reader.pages
+		for idx, page in enumerate(pages):
+			page_number = idx + 1
+			logger.log(
+				f"\tUploading blob for page {page_number} -> {os.path.splitext(target_file_name)[0]}-{page_number}.pdf"
+			)
+			# 1ページのPDFを切り出す
+			single_page_pdf = io.BytesIO()
+			writer = PdfWriter()
+			writer.add_page(page)
+			writer.write(single_page_pdf)
+			single_page_pdf.seek(0)
+			# blob storageにアップロード
+			container_client.upload_blob(
+				name=f"{os.path.splitext(target_file_name)[0]}-{page_number}.pdf",
+				data=single_page_pdf,
+				overwrite=True
+			)
+	# PDFファイル以外の場合
+	else:
+		print(f"Error: Must be a PDF file: {target_file_path}")
+		exit(1)
+
+
+# main実行関数
+def main():
+	# CLIから引数を取得
+	args, logger = read_argument()
+
+	logger.log("\nGetting credentials...")
+	# 認証方法を定義
+	default_creds = init_default_credential(
+		managedidentitycredential=args.managedidentitycredential,
+		searchkey=args.searchkey,
+		storagekey=args.storagekey,
+		tenantid=args.tenantid
+	)
+
+	# Search Serviceの認証情報を取得
+	search_creds = init_search_service_credential(
+		service_name=args.searchservice,
+		secret_key=args.searchkey,
+		default_creds=default_creds
+	)
+
+	# Form Recognizerの認証情報を取得
+	form_recognizer_creds = init_form_recognizer_credential(
+		service_name=args.formrecognizerservice,
+		secret_key=args.formrecognizerkey,
+		default_creds=default_creds
+	)
+
+	# Storage Accountの認証情報を取得
+	storage_creds = init_storage_credential(
+		storage_account_name=args.storageaccount,
+		secret_key=args.storagekey,
+		default_creds=default_creds
+	)
+
+	# Search Serviceのインデックスを作成
+	logger.log("Creating search index...")
+	create_search_service_index(search_creds, args.searchservice, args.index)
+
+	for pdfFile in glob(args.files):
+		# ファイルを読み込み、テキストを抽出
+		logger.log(f"\nExtracting text from {os.path.basename(pdfFile)}...")
+		document_content_list = recognize_text_from_pdf(
+			pdfFile,
+			args.formrecognizerservice, 
+			form_recognizer_creds
+		)
+
+		# ドキュメントをindexに追加
+		logger.log(f"Registering {os.path.basename(pdfFile)} to Search index...")
+		add_documents_to_index(
+			search_service_name=args.searchservice,
+			index_name=args.index,
+			search_creds=search_creds,
+			document_content_list=document_content_list,
+			logger=logger
+		)
+
+		# ドキュメントをblob storageにアップロード
+		logger.log(f"Uploading {os.path.basename(pdfFile)} to Blob Storage...")
+		upload_documents_to_blob_storage(
+			blob_service_name=args.storageaccount,
+			container_name=args.container,
+			target_file_path=pdfFile,
+			storage_creds=storage_creds,
+			logger=logger
+		)
+
+	logger.log("\nTask completed")
+
+main()
